Add fully-signed agreement workflow: PDF email + account storage
When both operator and Apex sign, automatically:
- Generate the final signed PDF with all signatures and initials
- Upload to Supabase storage (sales-documents bucket)
- Email the signed PDF to operator + james@apexaivending.com
- Save document record to operator's account (sales_documents)
- Update signed_pdf_url on the agreement

Extracted generatePurchaseAgreementPdf to shared lib so both
signing routes can generate PDFs without duplicating the 500-line
PDF builder.
